ALSA: Add EINTR handler to the polling function
Unity's garbage collector causes an EINTR error in the poll system call (maybe
due to the stop-the-world pause by GC). The current implementation of the input
stream polling function doesn't handle this error, and it stops the stream
thread when it happens. I changed it to resume polling after an EINTR error as
assuming it's harmless.